% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullsibQTL-internal.R
\name{calc_probs}
\alias{calc_probs}
\title{Function to obtain multipoint QTL probability for all chromossome   
it is used in create.fullsib}
\usage{
calc_probs(fullsib, step, error.prob, map.function, condIndex)
}
\description{
Function to obtain multipoint QTL probability for all chromossome   
it is used in create.fullsib
}
