\name{create_fullsib}

\alias{create_fullsib}
\alias{print.fullsib}

\title{Creates the main object to perform QTL mapping on \pkg{fullsibQTL}}

\description{
  First function that will be used on this package. It receives the
  data file containing all genotypes and phenotypes for all individuals
  on the mapping population. Also, requires the genetic map obtained
  from \pkg{onemap}. On this step, probabilities for QTL genotypes are
  obtained using Hidden Markov Models technology.
}

\usage{
create_fullsib(input.obj, map.list, step=0, error.prob=1e-4,
map.function=c("kosambi","haldane","morgan"), condIndex = 3.5) 

\method{print}{fullsib}(x, \ldots)

}

\arguments{
  \item{input.obj}{an object of class \emph{outcross}, obtained with
    \code{read.outcross} function from \pkg{onemap} package (>2.0-3), 
    where it is also possible to include phenotypic values.}
  
  \item{map.list}{A map obtained with \pkg{onemap}, it is either an object of
    class \emph{sequence} or list of objects from the same class.}
  
  \item{step}{Defines the distance in centiMorgans (cM) between
    positions on the linkage groups that genotype probabilities are
    obtained. By default, the probabilities is calculated only at the
    marker locus \sQuote{step=0}. If \sQuote{step=1}, for example, at
    every 1 cM a virtual marker will be created to evaluate if there is
    assocition with phenotipic values.} 
  
  \item{error.prob}{Probability of genotyping error.}
  
  \item{map.function}{Mapping function that will be used to obtain the
    probabilities. One can choose between \emph{kosambi}, \emph{haldane} or
    \emph{morgan}.}

  \item{condIndex}{threshold used to evaluate the level of
    informativeness for QTL genotypes probabilities. See details.}
  
  \item{x}{An object of class \emph{fullsib} generated by
    \code{create_fullsib} or \emph{fullsib_cofactors} created by
    \code{cof_selection} or \code{cof_definition}.}

  \item{\ldots}{Further arguments, passed to other methods. Currently
  ignored.}
}

\details{
  In a fullsib cross, a priori we have four genetic classes segregating
  on the progeny. However it is not always possible to identify four
  different classes. This happens because some regions on the genetic
  map may be saturated with molecular markers which are not fully
  informative. For example, if a given linkage group only have markers
  segregating as 1:1 fashion, we are able to identify two differents
  genotipic classes, resulting in only one genetic effect significative.
  This can also happens with markers segregating in 1:2:1 (only two
  of three effects are estimable) and 3:1 fashion, as well some this
  markers. When this situation is verified, colinearity on the QTL
  genotype probabilities is found and may lead some problems on the
  analysis.

  To avoid this kind of problems, we investigated for each position on
  the genome, the number of different genetic classes and if the any
  relationship between then. If four classes are avaliable, the model is
  runned as usual. On the other hand, we infer the estimable effects and
  use just them on the analysis. For details of how model works under
  colinearity, check \pkg{fullsibQTL} Vignette.

  The method print returns a text output summarizing the object of class
  \emph{fullsib} or \emph{fullsib_cofactors}. Briefly, it indicates the
  total number of linkage groups, markers and map total size. For each
  group, it is indicated, its size, number of marker on the group and
  the markers segregation found on the group. It is also showed the grid
  used to obtain multipoint probabilities on the genome. If the object
  is also from \emph{fullsib_cofactors}, it is indicated the trait that
  selection was done, as well the markers location. If a selected marker
  is not in the map, \code{NA} will be indicated on its corresponding
  linkage groups. 

  
}
\value{
  An object of class \emph{fullsib} is returned, which has the following
  components:
  
  \item{map}{List with objects of class \emph{sequence} obtained with
    \pkg{onemap}. Each position of the list contains a different linkage
    group.} 

  \item{unlinked.mkr}{Vector contaning markers that are not placed in
    the genetic map. The markers are represented by the indices as done
    with \pkg{onemap} package.}
  
  \item{pheno}{Matrix containing the phenotypes to be used for QTL
    mapping. Rows represents individuals and columns, the traits.}

  \item{probs}{List (of size equal to \code{map}) containing the
    multipoint probabilities and additional required information. Each
    position of the list has four components:

    \sQuote{cond.prob}: an array containing the multipoint
    probabilities of dimension [n.ind, n.locus, 4], where
    n.ind is the number of individuals and n.locus the total of position
    (on a linkage group) where the probabilities were considered
    
    \sQuote{newmap}: list containing three components: \code{dist}, the
    cumulate distance between all markers (including virtual markers);
    \code{phase}, linkage phases between them, \code{type}, markers
    segregation type.
    
    \sQuote{step} is the walking grid used to estimate the probabilities
    along the linkage group;

    \sQuote{colin} has information concern on the locus informativeness, i.e,
    \code{geno.class} indicates how many different genotypes means are
    estimable (2, 3 or 4) and \code{type} indicates what kind the
    colinearity one may have on these probs (0, 1, \ldots, 9). For
    further details for colinearity, see the \pkg{fullsibQTL} Vignette.}
}

\seealso{
  \code{onemap},

  \code{\link{example_QTLfullsib}}
}

\references{

  Wu, R., Ma, C.-X., Painter, I., Zeng, Z.-B. (2002a) Simultaneous
  maximum likelihood estimation of linkage and linkage phases in
  outcrossing species. \emph{Theoretical Population Biology} 61:
  349-363.
  
  Wu, R., Ma, C.-X., Wu, S. S., Zeng, Z.-B. (2002b). Linkage
  mapping of sex-specific differences. \emph{Genetical Research}
  79: 85-96
}


\examples{

  data(example_QTLfullsib)

  ##fullsib is main object to perform QTL maping

  fsib <- create_fullsib(example_QTLfullsib,
                         list(LG1_final, LG2_final, LG3_final, LG4_final),
                         step=1,map.function="kosambi",condIndex=3.5)
}

\keyword{IO}


\author{Rodrigo Gazaffi, \email{rgazaffi@gmail.com}}