% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullsibQTL-internal.R
\name{draw_map_qtl}
\alias{draw_map_qtl}
\title{Draw map}
\usage{
draw_map_qtl(
  map.list,
  horizontal = FALSE,
  names = FALSE,
  grid = FALSE,
  cex.mrk = 1,
  cex.grp = 0.75,
  trait
)
}
\description{
Draw map
}
