% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullsibQTL-internal.R
\name{get_phase}
\alias{get_phase}
\title{This function draw the QTL configuration in the linkage group}
\usage{
get_phase(model, alpha.p, lod.ap, alpha.q, lod.aq, probs)
}
\description{
This function draw the QTL configuration in the linkage group
}
