\name{get_segr}

\alias{get_segr}

\title{Indicates the segregation of mapped QTL}

\description{
  Indicates the QTL segregation for the result generated by both
  \code{im_char} and \code{cim_char}.
}

\usage{
get_segr(fschar, probs1 = 0.05, probs2 = 0.05) 
}

\arguments{
  \item{fschar}{An object from class \emph{fullsib_char}.}

  \item{probs1}{probability used to identify significative effects on
  hypothesis \eqn{H_{01}}, \eqn{H_{02}} and \eqn{H_{03}}. See details.}

  \item{probs2}{probability used to identify significative hypothesis
    for \eqn{H_{04}}, \eqn{H_{05}} and \eqn{H_{06}}. See details.}
}

\details{
  QTL mapping procedure considered here, supposes that QTL can have
  differents segregation patterns on the progeny. In a diploid species,
  one can find QTL showing 1:1:1:1, 1:2:1, 3:1 and 1:1 fashion. This
  segregations are function of the number and the magnitude of
  significative effects.

  To identify the segregation, a two step procedure is done. First, we
  test if additive effects in both parents and dominance are
  signficative (\eqn{H_{01}}, \eqn{H_{02}} and \eqn{H_{03}}), and we infer
  if they are significative based on probs1). If more than one QTL if
  significative we need to test if they are similar in magnitude (not
  considering the signal). Probs2 is used to define if \eqn{H_{04}},
  \eqn{H_{05}} and \eqn{H_{06}}, are significative. With the
  interpretation of this hypothesis the segregation is inferred. The
  details about these tests are can be found in Gazaffi et
  al. (submitted). For the situations which some genetic effects are not
  estimable,  the explanation of segregation is present in
  \code{fullsibQTL} Vignette.
}

\value{
  This function returns text output indicating the QTL segregation
}

\references{
  Gazaffi, R., Mollinari, M., Pastina, M.M., Margarido, G.R.A, Garcia,
  A.A.F. (submitted) A model for quantitative trait loci mapping,
  linkage phase and segregation pattern estimation for outcrossing
  species. \emph{Tree Genetics and Genomes}
}

\seealso{
  \code{\link[fullsibQTL]{im_char}},
  
  \code{\link[fullsibQTL]{cim_char}},
  
  \code{\link[fullsibQTL]{draw_phase}},

}

\examples{

  data(example_QTLfullsib)

  fullsib <- create_fullsib(example_QTLfullsib,
                            list(LG1_final, LG2_final, LG3_final, LG4_final),
                            step=0,map.function="kosambi",condIndex=3.5)


  ###############################################
  ## cofactor selection using BIC (n.ind = 300)
  cofs.fs <- cof_selection(fullsib, pheno.col=1, k = log(300),
                           selection=1) 

  \dontrun{
  cim1 <- cim_scan(cofs.fs, pheno.col=1, ws = 22, LOD= TRUE, icim=FALSE)
  summary(cim1)
  plot(cim1)
  }

  qtl.lg3 <- cim_char(cofs.fs, pheno.col=1, ws=22, lg=3, pos="M38")
  get_segr(qtl.lg3, probs1=0.05, probs2=0.05)

  qtl.lg4 <- cim_char(cofs.fs, pheno.col=1, ws=22, lg=4, pos="M52")
  get_segr(qtl.lg4)
}

\author{Rodrigo Gazaffi \email{rgazaffi@gmail.com}}

\keyword{utilities}
