% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullsib_cofactors.R
\name{plot.fullsib_cofactors}
\alias{plot.fullsib_cofactors}
\alias{plot.cof_selection}
\title{Plot \code{fullsib_cofactors} object}
\usage{
\method{plot}{fullsib_cofactors}(x, horizontal = FALSE, grid = FALSE,
  cex.mrk = 1, cex.grp = 0.75, ...)
}
\arguments{
\item{x}{An object from class \emph{fullsib_cofactors}.}

\item{horizontal}{if \code{TRUE}, indicates that the map should be plotted
horizontally. Default is \code{FALSE}.}

\item{grid}{if \code{TRUE}, displays a grid in the background. Default is
\code{FALSE}.}

\item{cex.mrk}{the magnification to be used for markers.}

\item{cex.grp}{the magnification to be used for group axis annotation.}

\item{...}{Further arguments, passed to other methods. Currently ignored.}
}
\value{
A plot of a map with the \emph{fullsib_cofactors} positions.
}
\description{
This function is based on \code{\link[onemap]{draw_map}} presented in \pkg{onemap}. It is
used to plot genetic map with the cofactors on map
}
\examples{
  data(example_QTLfullsib)

  fullsib <- create_fullsib(example_QTLfullsib,
                            list(LG1_final, LG2_final, LG3_final, LG4_final),
                            step=0,map.function="kosambi",condIndex=3.5)



  ###############################################
  ## cofactor selection using BIC (n.ind = 300)

  ### just using markers that are placed on linkage groups (default)
  cofs.fs <- cof_selection(fullsib, pheno.col=1, k = log(300), selection =1)
  cofs.fs
  plot(cofs.fs)
  plot(cofs.fs, horizontal = TRUE, grid = TRUE, cex.mrk = 2.5, cex.grp = 1)
  
}
\seealso{
\code{\link[onemap]{draw_map}},
\code{\link[fullsibQTL]{cof_selection}},
\code{\link[fullsibQTL]{cof_definition}}
}
\keyword{utilities}
