% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_phase.R
\name{plot_fullsib_phases}
\alias{plot_fullsib_phases}
\title{Plot phases}
\usage{
plot_fullsib_phases(x, CI = NULL, ...)
}
\arguments{
\item{x}{object or list of objects of class fullsib_phases}

\item{CI}{plots confidence interval using 1 LOD (LOD1) or 2 LODs (LOD2)}

\item{...}{currently ignored}
}
\description{
Plot phases
}
