% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fullsib.R
\name{print.fullsib}
\alias{print.fullsib}
\title{Print method for object class 'fullsib'}
\usage{
\method{print}{fullsib}(x, ...)
}
\arguments{
\item{x}{object of class fullsib}

\item{...}{currently ignored}
}
\description{
Print method for object class 'fullsib'
}
