% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_phase.R
\name{print.fullsib_phases}
\alias{print.fullsib_phases}
\title{Print phases}
\usage{
\method{print}{fullsib_phases}(x, ...)
}
\arguments{
\item{x}{object of class fullsib_phases}

\item{...}{currently ignored}
}
\description{
Print phases
}
