% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullsib_scan.R
\name{print.fullsib_scan}
\alias{print.fullsib_scan}
\title{Prints object of class \emph{fullsib_scan}}
\usage{
\method{print}{fullsib_scan}(x, lg, pos, ...)
}
\arguments{
\item{x}{Object of class \emph{fullsib_scan} resulted from \code{im_scan} or
\code{cim_scan}.}

\item{lg}{Linkage group that will be printed. Default is to print all
scanned groups.}

\item{pos}{Position of the QTL, string(s).}

\item{\dots}{parameters to be passed.}
}
\value{
None
}
\description{
Prints the object of class \emph{fullsib_scan} created by \code{im_scan} or
\code{cim_scan}.
}
\examples{
  data(example_QTLfullsib)
  fullsib <- create_fullsib(example_QTLfullsib,
                            list(LG1_final, LG2_final, LG3_final, LG4_final), 
                            step = 0, map.function = "kosambi",
                            condIndex = 3.5)

  ## running im_scan:
  im1 <- im_scan(fullsib, lg = "all", pheno.col = 1, LOD = TRUE)

  im1
  print(im1, lg=2)
  print(im1, lg=2, pos="M27")
  print(im1, lg=c(1,4))

}
\seealso{
\code{\link[fullsibQTL]{im_scan}}
\code{\link[fullsibQTL]{cim_scan}}
\code{\link[fullsibQTL]{plot.fullsib_scan}}
\code{\link[fullsibQTL]{summary.fullsib_scan}}
\code{scanone} from \pkg{qtl} package
}
\author{
Rodrigo Gazaffi, \email{rgazaffi@gmail.com}
}
\keyword{utilities}
