\name{QTLexample}
\alias{QTLexample}
\alias{QTLfullsib.txt}
\alias{LG1.end}
\alias{LG2.end}
\alias{LG3.end}
\alias{LG4.end}
\alias{fullsib.data}

\docType{data}

\title{Simulated data set to perform QTL mapping in a full-sib cross}

\description{
  Loads some necessary variables to create the main object required for
  using \pkg{fullsibQTL}, using \code{create.fullsib}. The loaded
  objects are generated by \pkg{onemap} package.
}

\usage{data(QTLexample)}

\format{
  This data loads four objects of class \emph{sequence}
  (\dQuote{LG1.end}, \dQuote{LG2.end}, \dQuote{LG3.end} and
  \dQuote{LG4.end}), i.e., they are the ordered linkage group and the
  final result of \pkg{onemap} package).

  Also, on example directory of this package, one may find a example
  file \dQuote{QTLexample.txt} which contains 65 markers with several
  segregation pattern and two traits (\dQuote{pheno1} and
  \dQuote{pheno2}) for 300 individuals.

  If the file \dQuote{QTLexample.txt} was read with function
  \code{read.outcross} from \pkg{onemap} package (>=2.0.3) the object
  \dQuote{fullsib.data} will be created.
}

\details{
  To exemplify the functions present in this package, it is considered a
  simulation done in the article of Gazaffi et al. (submitted).
  
  In a summary, \dQuote{QTLfullsib.txt} has 65 markers, with different
  segregation patterns: 15 of type A.1, 1 A.4, 5 B1.5, 6 B2.6, 6 B3.7, 11
  C.8, 10 D1.9, 10 D2.14, 1 D2.18. The codification of the segregations
  fashion follows the notation of Wu et al. (2002a). Those markers are
  distributed in four linkage groups, in each one having 15 equally
  distributed with 10 cM and five more unlinked markers that were
  simulated to demonstrate some options specially for cofactors
  selection. The population size is 300 individuals.

  The genetic map represented by these markers can be viewed in the
  objects \dQuote{LG1.end}, \dQuote{LG2.end}, \dQuote{LG3.end} and
  \dQuote{LG4.end}, that are objects of class \emph{sequence}.

  There are two simulated phenotypes that are present in
  \dQuote{QTLexample.txt} file. \dQuote{pheno1} is the original
  phenotype with 300 individuals and the trait heritability is
  0.70. \dQuote{pheno2} is the same trait as pheno1, but with an
  extra source of variation, modelled by addcovar parameter, along the
  functions to perform QTL mapping.
}

\references{
  Gazaffi, R., Mollinari, M., Pastina, M.M., Margarido, G.R.A, Garcia,
  A.A.F. (submitted) QTL mapping model for characterizing segregation
  pattern and linkage phase for outcrossing species. \emph{Tree Genetics
  and Genomes}
  
  Margarido, G. R. A., Souza, A.P. and Garcia, A. A. F. (2007)
  OneMap: software for genetic mapping in outcrossing species.
  \emph{Hereditas} 144: 78-79.

  Wu, R., Ma, C.-X., Painter, I. and Zeng, Z.-B. (2002a)
  Simultaneous maximum likelihood estimation of linkage and linkage
  phases in outcrossing species. \emph{Theoretical Population Biology}
  61: 349-363.

  Wu, R., Ma, C.-X., Wu, S. S. and Zeng, Z.-B. (2002b). Linkage
  mapping of sex-specific differences. \emph{Genetical Research} 79:
  85-96.
}
  
\examples{
  data(QTLexample)
  ls()

  fullsib <- create.fullsib(fullsib.data,
                            list(LG1.end, LG2.end, LG3.end, LG4.end),
                            step=0,map.function="kosambi",condIndex=3.5)

  ##im analysis
  im1 <- im.scan(fullsib, pheno.col=1)

  \dontrun{
  plot(im1)
  summary(im1)
  print(im1, lg=4)
  }

  qtl1 <- im.char(fullsib, pheno.col=1, lg=1, pos="M1")

  ##cofactor definition
  cofs <- matrix(c("1","M1",
                   "2","M27",
                   "3","M37",
                   "4","M52"),4,2, byrow=TRUE)
  (cof.def <- cof.definition(fullsib, pheno.col=1, cof.pos=cofs))


  ##cofactor selection
  cofs.fs <- cof.selection(fullsib, pheno.col = 1, k = log(300),
  selection = 1)
  cofs.fs
  plot(cofs.fs)

  ##cim analysis
  cim1 <- cim.scan(cofs.fs, pheno.col=1, ws = 22, LOD= TRUE, icim=FALSE)
  \dontrun{
  plot(cim1)
  summary(cim1)
  print(cim1, lg=4)
  }

  qtl.lg3 <- cim.char(cofs.fs, pheno.col=1, ws=22, lg=3, pos="M38")
  draw.phase(cofs.fs, qtl.lg3, probs=0.05)
  get.segr(qtl.lg3)

  qtl.lg4 <- cim.char(cofs.fs, pheno.col=1, ws=22, lg=4, pos="M52")
  draw.phase(cofs.fs, qtl.lg4, probs=0.05)
  get.segr(qtl.lg4)

  r2.ls(fullsib, pheno.col=1, lg=c(3,4), pos=c("M38","M52"),
  ls.estimation=TRUE)
}

\seealso{
  \code{\link[fullsibQTL]{create.fullsib}},
  \code{onemap}

}

\keyword{datasets}


\author{Rodrigo Gazaffi, \email{rgazaffi@gmail.com}}
