\name{summary.fullsib.perm}
\alias{summary.fullsib.perm}
\title{Returns the threshold obtained with permutation test}

\description{
  Gives the Threshold obtained with permutation test. It is used for
  both \code{im.scan} and \code{cim.scan}, when \sQuote{n.perm} argument
  is higher than 0.
}

\usage{
\method{summary}{fullsib.perm}(object, alpha = c(0.05, 0.10), \ldots)
}

\arguments{
  \item{object}{An object of class \emph{fullsib.perm}, output of \code{im.scan}
    and \code{cim.scan}, when permutation test is performed.}
  
  \item{alpha}{Genome-wide significance levels.}

  \item{\ldots}{Further arguments, passed to other methods. Currently
    ignored.}

}

\value{
  It returns a matrix summarizing the results generated by permutation
  test. In the rows each level of probabiliy defined by user is showed,
  by default two rows is printed (alpha at 0.05 and 0.10). The
  number of columns indicates the ordered highest peaks. If threshold
  value obtained for column of \sQuote{peak.1} is considered, one is
  using the approach proposed by Churchill and Doerge (1994). If
  \code{peaks.2} is used, it corresponds to the approach proposed by
  Chen and Storey (2006).
}

\details{
  It receives an object of class \emph{fullsib.perm} and apply for each
  column a quantile function to calculate the thresholds, using default
  of function \code{quantile} from \code{stats} package.
}

\references{
  Chen, L., Storey, J.D. (2006) Relaxed Significance Criteria for
  Linkage Analysis. Genetics 173: 2371-2381
  
  Churchill, G.A., Doerge, R.W. (1994) Empirical Threshold Values for
  Quantitative Trait Mapping. Genetics 138: 963-971.
}

\seealso{
  \code{\link[fullsibQTL]{im.scan}},
  
  \code{\link[fullsibQTL]{cim.scan}},

  \code{\link[fullsibQTL]{plot.fullsib.perm}},
 

}

\examples{
  \dontrun{
  data(QTLexample)

  fullsib <- create.fullsib(fullsib.data,
                            list(LG1.end, LG2.end, LG3.end, LG4.end), 
                            step=0, map.function="kosambi", condIndex = 3.5) 

  im.perm <- im.scan(fullsib, pheno.col=1, LOD=TRUE, n.perm=1000,
                     write.perm="perm_values.txt") 
  summary(im.perm)
  summary(im.perm, alpha = 0.05)
  ## first peak: churchill and doerge (1994)
  ## second peak: chen and storey (2006)
  }

}

\author{Rodrigo Gazaffi \email{rgazaffi@gmail.com}}

\keyword{utilities}
