% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullsib_perm.R
\name{summary.fullsib_perm}
\alias{summary.fullsib_perm}
\title{Returns the threshold obtained with permutation test}
\usage{
\method{summary}{fullsib_perm}(object, alpha = c(0.05, 0.1), verbose = TRUE, ...)
}
\arguments{
\item{object}{An object of class \emph{fullsib_perm}, output of
\code{im_scan} and \code{cim_scan}, when permutation test is performed.}

\item{alpha}{Genome-wide significance levels.}

\item{verbose}{if TRUE print the heading information}

\item{\dots}{parameters to be passed.}
}
\value{
It returns a matrix summarizing the results generated by permutation
test. In the rows each level of probabiliy defined by user is showed, by
default two rows is printed (alpha at 0.05 and 0.10). The number of columns
indicates the ordered highest peaks. If threshold value obtained for column
of \sQuote{peak.1} is considered, one is using the approach proposed by
Churchill and Doerge (1994). If \code{peaks.2} is used, it corresponds to
the approach proposed by Chen and Storey (2006).
}
\description{
Gives the Threshold obtained with permutation test. It is used for both
\code{im_scan} and \code{cim_scan}, when \sQuote{n.perm} argument is higher
than 0.
}
\details{
It receives an object of class \emph{fullsib_perm} and apply for each column
a quantile function to calculate the thresholds, using default of function
\code{quantile} from \code{stats} package.
}
\examples{

  \dontrun{
  data( "example_QTLfullsib" )

  fullsib <- create_fullsib( example_QTLfullsib,
                             list( LG1_final, LG2_final, LG3_final, LG4_final ), 
                             step = 0, map.function = "kosambi", condIndex = 3.5 ) 

  im_perm <- im_scan( example_QTLfullsib, pheno.col = 1, LOD = TRUE, n.perm = 1000,
                      write.perm = "perm_values.txt" ) 
  summary( im_perm )
  summary( im_perm, alpha = 0.05 )
  ## first peak: churchill and doerge (1994)
  ## second peak: chen and storey (2006)
  
  summary( im_perm, alpha = 0.05, verbose = FALSE )
  }

}
\references{
Chen, L., Storey, J.D. (2006) Relaxed Significance Criteria for
Linkage Analysis. Genetics 173: 2371-2381

Churchill, G.A., Doerge, R.W. (1994) Empirical Threshold Values for
Quantitative Trait Mapping. Genetics 138: 963-971
}
\seealso{
\code{\link[fullsibQTL]{im_scan}}
\code{\link[fullsibQTL]{cim_scan}}
\code{\link[fullsibQTL]{plot.fullsib_perm}}
}
\author{
Rodrigo Gazaffi \email{rgazaffi@gmail.com}
}
\keyword{utilities}
