% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fullsibQTL.R
\name{tidy_fullsibQTL}
\alias{tidy_fullsibQTL}
\title{Merges fullsibQTL data in a single and tidy data table}
\usage{
tidy_fullsibQTL(fullsib, fullsib.scan, r2ls.out = NULL, qtlmapping = NULL)
}
\arguments{
\item{fullsib}{An object from class \code{fullsib}.}

\item{fullsib.scan}{A list of object(s) from class \code{fullsib_scan}.}

\item{r2ls.out}{A list of object(s) from class \code{r2ls_out}.}

\item{qtlmapping}{character vector with the name(s) of \code{fullsib.scan}
objects to be ploted.  If \code{qtlmapping=NULL} plot generic names.}
}
\value{
A merged \code{data.table} with the input data
}
\description{
Merges in a single and tidy \code{data.table} objects from one or more of 
following classes: \code{fullsib}, \code{fullsib_scan}, and \code{r2ls_out}.
This function was developed to be used as part of the 
\code{\link[fullsibQTL]{plot_fullsibQTL}}.
}
\examples{

  library(fullsibQTL)
  data("example_QTLfullsib")
  fullsib <- create_fullsib(example_QTLfullsib,
                             list(LG1_final, LG2_final, LG3_final, LG4_final),
                             step=1,map.function="kosambi",condIndex=3.5)

  ## IM
  im <- im_scan(fullsib)
  qtls.im <- r2_ls(fullsib, pheno.col=1, lg=c(1,1,2,2,2,3,3,4),
                 pos=c("loc13","M12","loc45","loc77","loc108","M33","loc65","M52"))
  ## CIM
  cofs.fs <- cof_selection(fullsib, pheno.col=1, k=log(300), n.cofactor = 10)
  cim<-cim_scan(cofs.fs)
  qtls.cim <- r2_ls(fullsib, pheno.col=1, lg=c(1,1,2,2,2,3,3,4),
                    pos=c("loc13","M12","loc45","loc77","loc108","M33","loc65","M52"))

  ## Merging and transforming to tidy data
  tidy_fullsibQTL(fullsib, fullsib.scan=list(im,cim),
                  r2ls.out= list(qtls.im,qtls.cim), qtlmapping=c("im","cim"))
                  
}
\seealso{
\code{\link[fullsibQTL]{create_fullsib}}
\code{\link[fullsibQTL]{im_scan}}
\code{\link[fullsibQTL]{cim_scan}}
\code{\link[fullsibQTL]{r2_ls}}
\code{\link[fullsibQTL]{tidy_fullsibQTL}}
\code{\link[fullsibQTL]{plot_fullsibQTL}}
\code{\link[ggplot2]{ggplot}}
}
\author{
Rodrigo Amadeu, \email{rramadeu@gmail.com}
}
\keyword{utilities}
